fix error in live server

+++ lib/modules/bus_new/model/new_bus_search_list_response.dart
@@ -46,7 +46,7 @@ class Buses {
   // List<dynamic>? imgList;
   List<String>? amenities;
   // List<dynamic>? detailImage;
-  int? ticketPrice;
+  double? ticketPrice;
   // List<dynamic>? passengerDetail;
   List<SeatLayout>? seatLayout;
   String? operatorName;
